[COR-3748] updated amounts to int64

+++ pkg/moov/underwriting_v2507_models.go
@@ -71,8 +71,8 @@ type VolumeShareByCustomerType struct {
 }
 
 type EstimatedActivity struct {
-	AverageTransactionAmount *int                `json:"averageTransactionAmount,omitempty"`
-	MaximumTransactionAmount *int                `json:"maximumTransactionAmount,omitempty"`
+	AverageTransactionAmount *int64              `json:"averageTransactionAmount,omitempty"`
+	MaximumTransactionAmount *int64              `json:"maximumTransactionAmount,omitempty"`
 	MonthlyVolumeRange       *MonthlyVolumeRange `json:"monthlyVolumeRange,omitempty"`
 }
